Add an availability check URL for applications
Add a URL for applications to register to allow for per-app availability
checking of sources.

This will be invoked by the normal availabililty check (scheduled and
on-demand) for all enabled applications for a source.

Applications will receive a POST action to check the availability of all
related resources for the specified source by ID, then the application
will update the source via the API to set the availability_status.