Don't return the 'tenant' attribute
The tenant attribute maps to the user's identity.account_number which is
only relevant internally.  Users of the API who login through 3scale
will only ever be able to see a single tenant so this attribute isn't
helpful and can be confusing.  Also since it isn't a proper relationship
(it is the tenant.external_tenant not tenant_id) it can trip up things
like filtering.  In v2.0 we should drop this attribute.

Relates to JIRA: TPINVTRY-742